+++ co/buffer.h
@@ -46,9 +46,9 @@ class Buffer : public lunchbox::Bufferb, public lunchbox::Referenced
 {
 public:
     /** Construct a new buffer. @version 1.0 */
-    CO_API Buffer( BufferListener* listener = 0 );
+    CO_API explicit Buffer( BufferListener* listener = 0 );
     
-    /** Construct a new buffer. @version 1.0 */
+    /** Construct a new buffer. @version 1.0-vmml */
     CO_API Buffer( lunchbox::Bufferb& from, BufferListener* listener = 0 );
 
     /** Destruct this buffer. @version 1.0 */